+++ src/constants/devOpsDisplayResources.ts
@@ -1,0 +1,10 @@
+export const baseUriEntryPrompt =
+  "Please enter your DevOps Base Url here. For example [https://dev.azure.com/OrganisationName/]...";
+export const tokenEntryPrompt = "Please enter your DevOps Personal Access Token here...";
+export const connectionFailed =
+  "Connection to the DevOps API failed with the provided authentication. Has your PAT expired or been entered incorrectly?";
+export const missingDetails = "Please authorize your requests first using the [DevOps: Authorize] command.";
+export const retrievingRepositories = "Retrieving repositories...";
+export const selectRepoToClone = "Select a repository to clone...";
+export const authorizationCleared = "Stored DevOps authorization has been cleared.";
+export const authorizationStored = "Authorization to DevOps has been validated and stored successfully.";