+++ src/constants/globalStateResources.ts
@@ -1,0 +1,2 @@
+export const devOpsBaseUri = "dwvs_devOpsBaseUri";
+export const devOpsPAT = "dwvs_devOpsPAT";