+++ src/contexts/developmentContext.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import { DisplayService } from "../services/displayService";
+import { GitExtensionService } from "../services/gitExtensionService";
 
 export class DevelopmentContext implements vscode.Disposable {
   //Contexts
@@ -7,11 +8,13 @@ export class DevelopmentContext implements vscode.Disposable {
 
   //Services
   displayService: DisplayService;
+  gitExtensionService: GitExtensionService;
 
   constructor(context: vscode.ExtensionContext) {
     this.vscodeExtensionContext = context;
 
     //Initialise services
+    this.gitExtensionService = new GitExtensionService(this);
     this.displayService = new DisplayService(this);
   }