+++ src/contexts/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from "vscode";
 import { ALDevelopmentContext } from "./alDevelopmentContext";
 import { DevelopmentContext } from "./developmentContext";
+import { DevOpsDevelopmentContext } from "./devopsDevelopmentContext";
 
 export function activate(context: vscode.ExtensionContext) {
   //Initialise contexts
@@ -9,6 +10,9 @@ export function activate(context: vscode.ExtensionContext) {
 
   const alDevelopmentContext: ALDevelopmentContext = new ALDevelopmentContext(context, developmentContext);
   context.subscriptions.push(alDevelopmentContext);
+
+  const devOpsDevelopmentContext: DevOpsDevelopmentContext = new DevOpsDevelopmentContext(context, developmentContext);
+  context.subscriptions.push(devOpsDevelopmentContext);
 }
 
 export function deactivate() {}